Rename 'constants' module -> 'transforms'.
In preparation for adding 'ArrayUnion' / 'ArrayRemove' transforms, which
are not constants.

See: #6551.
