--- firestore/docs/constants.rst
@@ -1,6 +1,0 @@
-Constants
-~~~~~~~~~
-
-.. automodule:: google.cloud.firestore_v1beta1.constants
-  :members:
-  :show-inheritance:
