+++ firestore/docs/transforms.rst
@@ -1,0 +1,6 @@
+Transforms
+~~~~~~~~~~
+
+.. automodule:: google.cloud.firestore_v1beta1.transforms
+  :members:
+  :show-inheritance:
